requirements.txt
# MCQ Extractor — Python dependencies
# Install with:  pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# PDF processing
PyMuPDF==1.24.2

# Image processing
Pillow==10.3.0
numpy==1.26.4

# LLM client (OpenAI-compatible, works with Nebius)
openai>=1.54.0
jinja2==3.1.2
httpx>=0.27.0,<1.0.0

# Database
pymongo==4.7.2

# Environment variable loading
python-dotenv==1.0.1
